refactor to use traits
