+++ api/app/Traits/Generators/GeneratesDoc.php
@@ -1,0 +1,24 @@
+<?php
+
+namespace App\Traits\Generator;
+
+use PhpOffice\PhpWord\Element\Section;
+
+trait GeneratesDoc
+{
+    use GeneratesFile;
+
+    /**
+     * Adds text accomponied by a strong label
+     *
+     * @param  Section  $section  The section to add it to
+     * @param  string  $label  Label for the text
+     * @param  string  $string  The value
+     */
+    protected function addLabelText(Section $section, string $label, string $text)
+    {
+        $run = $section->addTextRun();
+        $run->addText($label.$this->colon());
+        $run->addText($text);
+    }
+}
